// src/shared/enums/FilterEnum.ts
export enum SortByEnum {
    popularityAsc = 'popularity.asc',
    popularityDesc = 'popularity.desc',
    revenueAsc = 'revenue.asc',
    revenueDesc = 'revenue.desc',
    primaryReleaseDateAsc = 'primary_release_date.asc',
    primaryReleaseDateDesc = 'primary_release_date.desc',
    voteAverageAsc = 'vote_average.asc',
    voteAverageDesc = 'vote_average.desc',
    voteCountAsc = 'vote_count.asc',
    voteCountDesc = 'vote_count.desc'
}